+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [TT-4418] Add support for create/edit scheduled trip endpoints
 * [TT-4433] Add network interceptor for CRSF token
 * [TT-4429] Fix routes product_type_id parameter
+* [TT-4435] Add support for checkout endpoint
 
 ## 0.5.0